Raw HTTP client (HttpClient), PSR7Client now depends on HttpClient
